Address second-pass PR review feedback
- WindowsAppDriver.ClickPoint: cast inputs.Length to uint when comparing
  against SendInput's uint return value so the Windows build compiles.
- AppDriverFactory: include 'wpf' in the ArgumentException message so the
  error reflects the actual accepted platform list.
- DevFlowAgentService.CaptureUiOrNativeAsync: add a single in-flight gate
  (_pendingCaptureUiTask) so repeated tree/query calls don't accumulate
  unbounded queued UI-dispatch work while the dispatcher is blocked.
  Subsequent callers fall straight through to a native-only probe until
  the previously queued uiTask drains.
- NativeWindowProbe: add ExtractHwndsFromId helper that parses the embedded
  HWND from a 'native:hwnd:0x..' element id.
- PlatformVisualTreeWalker and WpfVisualTreeWalker: on cache miss, re-walk
  the native tree seeded with the HWND parsed from the requested id so
  ':dialog:{n}' prefixes from AppendKnownWindowDialogSubtrees are
  regenerated and id resolution remains stable across calls.

+++ src/DevFlow/Microsoft.Maui.DevFlow.Driver/WindowsAppDriver.cs
@@ -720,7 +720,7 @@ public class WindowsAppDriver : AppDriverBase
             new() { type = INPUT_MOUSE, u = new INPUTUNION { mi = new MOUSEINPUT { dwFlags = MOUSEEVENTF_LEFTDOWN } } },
             new() { type = INPUT_MOUSE, u = new INPUTUNION { mi = new MOUSEINPUT { dwFlags = MOUSEEVENTF_LEFTUP } } }
         };
-        return SendInput((uint)inputs.Length, inputs, Marshal.SizeOf<INPUT>()) == inputs.Length;
+        return SendInput((uint)inputs.Length, inputs, Marshal.SizeOf<INPUT>()) == (uint)inputs.Length;
     }
 
     private static ushort MapKeyToVirtualKey(string key) => key.ToUpperInvariant() switch
